% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablanor_plot_instillinger.R
\name{ablanor_plot_theme}
\alias{ablanor_plot_theme}
\title{Add ablanor theme to ggplot chart}
\usage{
ablanor_plot_theme()
}
\description{
This function allows you to add the ablanor theme to your ggplotgraphics.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(x = hp, y = mpg, color=cyl, shape=cyl)) +
geom_point(size=3) +
ablanor::ablanor_plot_theme()
}
}
\keyword{ablanor_plot_theme}
