% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreger_ki_prop.R
\name{aggreger_ki_prop}
\alias{aggreger_ki_prop}
\title{Regn ut Kvalitetsindikator - Andel
Funksjon for å regne ut kvalitetsindikatorer for andeler. Tar inn et
datasett på 101-format, og returnerer et estimert resultat}
\usage{
aggreger_ki_prop(d_ki_ind, alfa = 0.05, forenklet = FALSE)
}
\arguments{
\item{d_ki_ind}{Inndata på 101-format}

\item{alfa}{Verdi for å bestemme bredde på konfidensintervall,
standardverid er 0.05}

\item{forenklet}{bools. FALSE gir estimatet bare etter testing av d_ki_ind.
TRUE gir estimatet uten testing.}
}
\value{
Sammendrag av estimert proporsjon med konfidensintervall.
}
\description{
Regn ut Kvalitetsindikator - Andel
Funksjon for å regne ut kvalitetsindikatorer for andeler. Tar inn et
datasett på 101-format, og returnerer et estimert resultat
}
\examples{
\dontrun{
aggreger_ki_prop(d_ki_ind, alfa = 0.05)
}
}
