% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{def_utledete_var}
\alias{def_utledete_var}
\title{Utledete variabler}
\format{
data frame with variables to describe "Utledete variabler".
\describe{
\item{ledetekst}{}
\item{fysisk_feltnavn}{}
\item{skjemanavn}{}
\item{listeverdier}{}
\item{listetekst}{}
}
}
\usage{
def_utledete_var
}
\description{
Utledete variabler
}
\keyword{datasets}
