% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_name.R
\name{get_hospital_name}
\alias{get_hospital_name}
\title{Get hospital name from db}
\usage{
get_hospital_name(resh_id)
}
\arguments{
\item{resh_id}{Integer resh id}
}
\value{
Character string with hospital name
}
\description{
Get hospital name from db based on resh id
}
