% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\alias{getProsPatient}
\alias{getRand12}
\title{Load data for ablanor}
\usage{
getRand12(registryName, singleRow, ...)

getProsPatient(registryName, singleRow, ...)
}
\arguments{
\item{registryName}{Character string defining the registry name}

\item{singleRow}{Logical if only one row from the table is to be provided.
Default value is FALSE}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame or (when multiple data sets are returned) a list of data
frames containing registry data
}
\description{
Functions for loading data needed by ablanor from a database or local files.
When local files are used the environmental variable
\emph{filbane_ablanor_test} must be set to the path of the file where data
are found
}
