% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDump.R
\name{getDataDump}
\alias{getDataDump}
\title{Get datadump ablanor}
\usage{
getDataDump(tableName, fromDate, toDate, reshId = NULL, userRole, ...)
}
\arguments{
\item{tableName}{Selected table for download}

\item{fromDate}{default 01-01-1900. Start date in calendar if
tableName is pros or basereg, argument ignored for other tableNames.}

\item{toDate}{default is newest registration. End date i calendar
if tableName is pros or basereg, argument ignored for other tableNames.}

\item{reshId}{from session information}

\item{userRole}{"SC" gets National data, "LC" get local hospital's data.
Some tables are only available for SC users.}
}
\value{
data.frame
}
\description{
Use getDataAbalnor function to get selected table.
}
