% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKodebokData.R
\name{getKodebokMedUtledetedVar}
\alias{getKodebokMedUtledetedVar}
\title{Legg til utledete variabler i kodeboken}
\usage{
getKodebokMedUtledetedVar()
}
\value{
data.frame med kodeboken til AblanNor med rader for utledete
variabler Nøkkelkolonner:  \code{fysisk_feltnavn} (variabel-navn)
\code{listeverdier} (numeriske verdier for Listevariabler),
\code{listetekst} (tekst-verdi for Listevariabler),
\code{skjemanavn} (opprinnelse av variabelen) og
\code{ledetekst}(forklaring av variabelen).
Hver variabel har en rad i \code{kodeboken}. For variabler av \code{type}
"Listevariabel"  inneholder kodeboken en rad per listeverdi.
}
\description{
Les inn kodeboken og legg til definisjoner på utledete variabler. Velge ut
kun enklelte kolonner. Brukes i "Kodebok"-fanen på Rapporteket.
}
