% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrepData.R
\name{getPrepDataAblanor}
\alias{getPrepDataAblanor}
\alias{getBaseregDatagetProsData}
\alias{getMceData}
\alias{getRand12Data}
\alias{getFollowupBasisData}
\alias{getFollowupOneYrData}
\alias{getFollowupFiveYrData}
\alias{getGkvData}
\alias{getPromsData}
\alias{getBaseregProsData}
\alias{getBaseregProsHendelseData}
\alias{getBaseregProsFollowup1Data}
\alias{getBaseregProsFollowup5Data}
\alias{getBaseregProsFollowup0Data_v2}
\alias{getBaseregData}
\alias{getProsData}
\alias{getBaseregProsFollowup1Data_v2}
\title{Data managment on tables}
\usage{
getBaseregData(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getProsData(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getMceData(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getRand12Data(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupBasisData(singleRow = FALSE, reshId = NULL, userRole, ...)

getFollowupOneYrData(singleRow = FALSE, reshId = NULL, userRole, ...)

getFollowupFiveYrData(singleRow = FALSE, reshId = NULL, userRole, ...)

getGkvData(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getPromsData(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsData(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsHendelseData(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsFollowup0Data_v2(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsFollowup1Data(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsFollowup1Data_v2(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsFollowup5Data(
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)
}
\arguments{
\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{reshId}{Integer organization id. From login settings.}

\item{userRole}{String dummy/placeholder role. "LC" has access only
to local data (defined by reshId), "SC" has access to national data.}

\item{fromDate}{NULL default is 01-01-1900. If datadump or pivot table,
start date of calendar is used.}

\item{toDate}{NULL default is newest registration in Abalnor. If datadump or
pivot table, end date of calendar is used.}
}
\value{
data.frame med rad per forløp og kolonner for variabler
}
\description{
Load data and apply data-management operations. Tables can be used in
\emph{pivot-tables} or monthly reports. Notice, only tables when a
procedure-date exists!
}
