% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProsPatientData.R
\name{getProsPatientData}
\alias{getProsPatientData}
\title{Hent merget datasett
Hent datasett fra basis-skjema, prosedyreskjema, mce, og pasientskjema. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til utledete
variabler og nye variabler.}
\usage{
getProsPatientData(
  registryName,
  singleRow = FALSE,
  reshID = NULL,
  allData = FALSE,
  ...
)
}
\arguments{
\item{registryName}{Dersom verdien "test_ablanor_lokalt" leser vi inn
lokal RData-fil. Ellers er det SQL spørring}

\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{allData}{Logical if global data set is to be returned. When FALSE
(default) data will be filtered by \code{reshId}}

\item{...}{Optional arguments to be passed to the function}

\item{reshId}{Integer organization id}
}
\value{
data.frame med rad per forløp og kolonner for variabler
}
\description{
Hent merget datasett
Hent datasett fra basis-skjema, prosedyreskjema, mce, og pasientskjema. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til utledete
variabler og nye variabler.
}
