% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kbd}
\alias{kbd}
\title{Klokeboken data dictionary}
\format{
Data frame containing Klokeboken data
\describe{
\item{skjemanavn}{}
\item{navn_i_rapporteket}{}
\item{ledetekst}{}
\item{obligatorisk}{}
\item{type}{}
\item{listeverdier}{}
\item{listetekst}{}
\item{normalintervall_start_numerisk}{}
\item{normalintervall_slutt_numerisk}{}
\item{maksintervall_start_numerisk}{}
\item{maksintervall_slutt_numerisk}{}
\item{normalintervall_start_dato}{}
\item{normalintervall_slutt_dato}{}
\item{maksintervall_start_dato}{}
\item{maksintervall_slutt_dato}{}
\item{antall_tegn}{}
\item{lovlige_tegn}{}
\item{desimaler}{}
\item{aktiveringsspoersmaal}{}
\item{underspoersmaal}{}
\item{innfoert_dato}{}
\item{utfaset_dato}{}
\item{tabell}{}
\item{fysisk_feltnavn}{}
\item{kommentar}{}
\item{variabel_id}{}
\item{hjelpetekst}{}
}
}
\usage{
kbd
}
\description{
Hard copy of Ablanor Klokeboken
}
\keyword{datasets}
