% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorar.R
\name{ki_akutt_suksess}
\alias{ki_akutt_suksess}
\alias{ki_akutt_suksess_svt}
\alias{ki_akutt_suksess_svt_avrt}
\alias{ki_akutt_suksess_svt_avnrt}
\alias{ki_akutt_suksess_svt_ikkje_avrt_el_avnrt}
\title{Kvalitetsindikatorar for akutt suksess
Lag KI-datasett for om ein prosedyre skal reknast som akutt suksess (ja/nei).}
\usage{
ki_akutt_suksess(df)

ki_akutt_suksess_svt(df)

ki_akutt_suksess_svt_avrt(df)

ki_akutt_suksess_svt_avnrt(df)

ki_akutt_suksess_svt_ikkje_avrt_el_avnrt(df)
}
\arguments{
\item{df}{Prosedyredatasett.
Viss variabelen for akutt suksess har verdien «usikker» (2), vert det rekna
som «nei». Berre prosedyrarar der pasienten faktisk vart abladert er med i
utrekninga (observasjonar med verdien «ikke aktuelt» (3) for akutt suksess
inngår altså ikkje).

Det finst òg nokre spesialfunksjonar for undergrupper:
\itemize{
\item \code{ki_akutt_suksess_svt()}: Ser berre på SVT-forløp.
\item \code{ki_akutt_suksess_svt_avrt()}: Ser berre på SVT-forløp med undergruppa
AVRT (aksessoriske baner).
\item \code{ki_akutt_suksess_svt_avnrt()}: Ser berre på SVT-forløp med undergruppa
AVNRT (AV nodal reentry).
\item \code{ki_akutt_suksess_svt_ikkje_avrt_el_avnrt()}: Ser berre på SVT-forløp,
men ekskluderer undergruppene AVRT og AVNRT.
}}
}
\value{
KI-datasett eigna for bruk med \code{\link[=aggreger_ki_prop]{aggreger_ki_prop()}}.
}
\description{
Kvalitetsindikatorar for akutt suksess
Lag KI-datasett for om ein prosedyre skal reknast som akutt suksess (ja/nei).
}
