% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorar.R
\name{ki_dod}
\alias{ki_dod}
\title{Kvalitetsindikator for død
Lag KI-datasett for om pasienten døde innanfor eit
visst tal dagar frå prosedyren.}
\usage{
ki_dod(d_pros, d_mce, d_pas, dagar_innan = 29)
}
\arguments{
\item{d_pros}{Prosedyredatasett.}

\item{d_mce}{Forløpsdatasett.}

\item{d_pas}{Pasientdatasett (\code{patientlist}-tabellen).}

\item{dagar_innan}{Talet på dagar som ein skal sjå om
pasienten døde innan. Dødsfall \emph{på} denne dagen
vert òg rekna med. Standardverdien er 29, dvs. ein
ser om pasienten døde < 30 dagar frå prosedyredatoen.}
}
\value{
KI-datasett eigna for bruk med \code{\link[=aggreger_ki_prop]{aggreger_ki_prop()}}.
}
\description{
Kvalitetsindikator for død
Lag KI-datasett for om pasienten døde innanfor eit
visst tal dagar frå prosedyren.
}
\details{
Berre prosedyrar gjort minst \code{dagar_innan}
før \emph{dagens dato} inngår i indikatoren, pluss
prosedyrar der pasienten døde under operasjon.
Grunnen er at det berre er desse me garantert kan avgjera
om pasienten overlevde eller ikkje.
}
\note{
Indikatoren kan ha misvisande tal (for mange overlevande)
for prosedyrar veldig nær datadumpdatoen, spesielt for historiske
datadumpar. I framtida vil funksjonen kanskje
endrast til å ta inn ein eksplisitt dato
for når me har oppdatert dødsdatoinformasjon frå.
}
