% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorar.R
\name{ki_komplikasjonar}
\alias{ki_komplikasjonar}
\title{Kvalitetsindikator
Kvalitetsindikator for komplikasjoner}
\usage{
ki_komplikasjonar(df)
}
\arguments{
\item{df}{data frame med kolonnen komp_janei}
}
\value{
data frame med to nye kolonner : ki_krit_teller og ki_krit_nevner.
Dette KI-datasettet kan bli brukt med \code{\link[=aggreger_ki_prop]{aggreger_ki_prop()}}
}
\description{
Kvalitetsindikator
Kvalitetsindikator for komplikasjoner
}
