% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorar.R
\name{ki_komplikasjonar_pacemaker}
\alias{ki_komplikasjonar_pacemaker}
\title{Kvalitetsindikator for behov for pacemaker
Lag KI-datasett for prosedyrerelatert AV-blokk med behov
for etterfølgjande pacemakerimplantasjon (ja/nei).}
\usage{
ki_komplikasjonar_pacemaker(df)
}
\arguments{
\item{df}{Prosedyredatasett.}
}
\value{
KI-datasett eigna for bruk med \code{\link[=aggreger_ki_prop]{aggreger_ki_prop()}}.
}
\description{
Kvalitetsindikator for behov for pacemaker
Lag KI-datasett for prosedyrerelatert AV-blokk med behov
for etterfølgjande pacemakerimplantasjon (ja/nei).
}
