% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorar.R
\name{ki_komplikasjonar_tamponade}
\alias{ki_komplikasjonar_tamponade}
\title{Kvalitetsindikator for tamponade
Lag KI-datasett for tamponade som følge av operasjon
som fører til intervensjon eller forlenget sykehusopphold.}
\usage{
ki_komplikasjonar_tamponade(df)
}
\arguments{
\item{df}{Prosedyredatasett.}
}
\value{
KI-datasett egnet for bruk med \code{\link[=aggreger_ki_prop]{aggreger_ki_prop()}}.
}
\description{
Kvalitetsindikator for tamponade
Lag KI-datasett for tamponade som følge av operasjon
som fører til intervensjon eller forlenget sykehusopphold.
}
