% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kodebok_funksjoner.R
\name{kodebok_funksjoner}
\alias{kodebok_funksjoner}
\alias{kodebok_sjekk_foer_leggtil}
\alias{kodebok_sjekk_foer_fjerning}
\alias{kodebok_fyll_listetekstvar}
\alias{kodebok_beholde_bare_listetekstvar}
\title{AblaNor kodebok functions}
\usage{
kodebok_sjekk_foer_leggtil(df, verdi_variabel, tekst_variabel, koder)

kodebok_sjekk_foer_fjerning(df, kb, verdi_variabel, tekst_variabel)

kodebok_fyll_listetekstvar(df, kb, ..., suffiks = "_tekst")

kodebok_beholde_bare_listetekstvar(
  df,
  kb,
  ...,
  suffiks = "_tekst",
  fjerne_suffiks_fra_navn = TRUE
)
}
\arguments{
\item{df}{data.frame containing variables from AblaNor e.g. \code{d_ablanor}.}

\item{verdi_variabel}{is used to check compatibility between \code{df} and \code{kb}
for one variable. \code{verdi_variabel} is the name of variable with numerical values
(\emph{listeverdier}).}

\item{tekst_variabel}{is used to check compatibility between \code{df} and \code{kb}
for one variable. \code{tekst_variabel} is the name of variable with text values
(\emph{listetekst}).}

\item{koder}{is used to check compatibility between \code{df} and \code{kb}
for one variable. \code{koder} contains selected rows from \code{kb}.}

\item{kb}{data.frame with \code{kodeboken}. Must contain variables
\code{type}, \code{fysisk_feltnavn}, \code{listeverdier} and
\code{listetekst}. One row for each level of all variables.}

\item{...}{If only selected \emph{listetekst}- or \emph{listeverdi}-
variables are to be added or removed in \code{df}, the selected variables
should be listed here, separated by commas.}

\item{suffiks}{string value that contains suffix of new \emph{listetekst}-
variables to be added to or \emph{listeverdi}-variables to be removed from
\code{df}. Default value is \emph{"_tekst"}.}

\item{fjerne_suffiks_fra_navn}{boolean. If TRUE the \emph{listetekst}-
variables are renamed and suffix is removed from variable names. If FALSE,
variable-names are not changed and will contain suffix.}
}
\description{
Functions for adding and removing \code{listeverdi}- and \code{listetekst}-
variables in \emph{AblaNor}-data. Functions are:
\itemize{
\item kodebok_sjekk_foer_leggtil() checks compatibility between \code{df}
and \code{kb} for a given variable \code{verdi_variabel} and says whether
corresponding \code{tekst_variabel} can be added in \code{df}.
\item kodebok_sjekk_foer_fjerning() checks compatibility between \code{df}
and \code{kb} for a given variable \code{tekst_variabel} and says and whether
corresponding \code{verdi_variabel} can be removed from \code{df}.
\item kodebok_fyll_listetekstvar() adds some/all \emph{listetekst}-variables
in \code{df}.
\item kodebok_beholde_bare_listetekstvar() removes some/all
\emph{listeverdi}-variables from \code{df}.
}
}
\examples{
 df <- data.frame(var1 = c(1:5, 1:5),
                  var2 = rep(c(0,1), 5))
 kb <- data.frame(fysisk_feltnavn = c(rep("var1", 5), "var2", "var2"),
                  type = rep("Listevariabel", 7),
                  listeverdier = c(1:5, 0, 1),
                  listetekst = c(letters[1:5], "nei", "ja"))

 ablanor::kodebok_sjekk_foer_leggtil(
                         df = df,
                         tekst_variabel = "var2_tekst",
                         verdi_variabel = "var2",
                         koder = kb \%>\%
                           dplyr::filter(.data$fysisk_feltnavn == "var2"))

ablanor::kodebok_fyll_listetekstvar(df = df,
                                    kb = kb,
                                    suffiks = "_tekst")
df <- data.frame(var1 = c(1:5),
                 var1_tekst = letters[1:5],
                 var3 = c(0, 0, 1, 1, 1),
                 var3_tekst = c("nei", "nei", "ja", "ja", "ja"))

kb <- data.frame(fysisk_feltnavn = c(rep("var1", 5), "var3", "var3"),
                 type = rep("Listevariabel", 7),
                 listeverdier = c(1:5, 0, 1),
                 listetekst = c(letters[1:5],"nei", "ja"))
ablanor::kodebok_sjekk_foer_fjerning(df = df,
                                     kb,
                                     verdi_variabel = "var3",
                                     tekst_variabel = "var3_tekst")
ablanor::kodebok_beholde_bare_listetekstvar(
                           df = df,
                           kb = kb,
                           suffiks = "_tekst",
                           fjerne_suffiks_fra_navn = TRUE)
}
