% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvalitetsindikatorer_ny.R
\name{kvalitetsindikatorer_ny}
\alias{kvalitetsindikatorer_ny}
\alias{ki_komplikasjoner}
\alias{ki_dod_ny}
\title{Ablanor sine kvalitetsindikatorer}
\usage{
ki_komplikasjoner(df)

ki_dod_ny(df, dager_innen = 30, dato_sensur)
}
\arguments{
\item{df}{Ablanors koblete tabell (pros_patient_followup). Må inneholde
utvalgte variabler, avhengeig av hvilke indikator som skal regnes ut. F.eks.
\code{ki_komplikasjoner()} trenger \emph{komp_janei} og \emph{pros_status}.}

\item{dager_innen}{Numerisk. Default verdi er 30. Brukes som parameter i
\code{ki_dod_ny()}, dersom antall dager mellom prosedyre og dødsfall er
under dette sifferet så vil forløpet telle i indikatoren.}

\item{dato_sensur}{Dato. Sensur-dato for overlevelse. Brukes som parameter i
\code{ki_dod_ny()} Det tar inntil 30 dager for å overføre informasjon om død
fra folkeregisteret til databasen. Forløp for nær dato for datadump bør ikke
være med i datagrunnlaget for indikatoren.}
}
\description{
Funksjoner som lager 2 nye variabler for hver av kvalitetsindikatorene: Den
første variabelen forteller hvorvidt forløpet er med i nevneren
(datagrunnlaget) for den aktuelle indikatorer, den andre variabelen
forteller om forløpet er med i telleren (ja, nei, NA). Telleren er alltid
NA dersom forløpet ikke er med i datagrunnlaget. Nevneren har suffix
\emph{_data}.
}
\details{
Følgende variabler er laget for de ulike funksjonene:

\code{ki_komplikasjoner()}
\itemize{
\item nevneren \code{indikator_komplikasjon_data} (datagrunnlag) er \emph{ja}
for alle ferdigstilte skjema og \emph{nei} for ikke-ferdigstilte skjema.
\item telleren \code{indikator_komplikasjon} er \emph{ja} dersom
\code{komp_janei} er lik 1 og \emph{nei} dersom \code{komp_janei} er lik 0.
}

\code{ki_dod_ny()}
\itemize{
\item nevneren \code{indikator_dod_data} (datagrunnlag) er \emph{nei}
for avdøde pasienter der enten \code{dato_pros} eller \code{deceased_date}
mangler, og \emph{nei} for levende pasienter der \code{dato_pros} mangler
eller det er færre enn \code{dager_innen} dager mellom prosedyren og
sensurdatoen (dette gjelder forløp nær datadump-dato, der eventuell
informasjon om død ikke har fått tid til å bli overført fra
folkeregisteret).
andre forløp er \code{indikator_dod_data} lik \emph{ja}.
\item telleren \code{indikator_dod} er \emph{ja} dersom pasienten er død innen
\code{dager_inne} dager etter prosedyren. \code{indikator_dod} er \emph{nei}
dersom pasienten ikke er død, eller er død over \code{dager_innen} (ofte 30)
dager etter prosedyren. Dersom forløpet ikke er med i datagrunnlaget, så
er \code{indikator_dod} lik \emph{NA}.
}
}
\examples{
x <- data.frame(
        mceid = 1:5,
        pros_status = c(-1, 0, 1, 1, 1),
        komp_janei = c(NA, 0, 1, 1, 1))
ablanor::ki_komplikasjoner(df = x)
x <- data.frame(
  dato_pros = as.Date(c("2021-01-19",
                        "2020-12-16",
                        "2020-12-16"), format = "\%Y-\%m-\%d"),
  deceased = c(0, 1, 1),
  deceased_date = as.Date(c(NA, "2022-01-12", "2020-12-16"),
                          format = "\%Y-\%m-\%d"))
ablanor::ki_dod_ny(df = x,
                   dato_sensur = as.Date("2022-01-20", format ="\%Y-\%m-\%d"))
}
