% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg_til_sykehusnavn.R
\name{legg_til_sykehusnavn}
\alias{legg_til_sykehusnavn}
\title{Add hospital name}
\usage{
legg_til_sykehusnavn(df, short = TRUE)
}
\arguments{
\item{df}{data.frame with AblaNor data. Must contain \code{centreid}.}

\item{short}{boolean. Short = TRUE gives short names, short = FALSE gives
complete names.}
}
\value{
Returns \code{df} with new one new column.
}
\description{
Add hospital name as text string based on value of \code{centreid}. New
variable is called \code{sykehusnavn}. \code{sykehusnavn} will overwrite
exising variable of same name.
}
\examples{
df <- data.frame(centreid = c(102966, 104284, 4214492, 700328, 700728))
ablanor::legg_til_sykehusnavn(df = df, short = TRUE)
}
