% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leggTilTommePivotWide.R
\name{leggTilTomme_PivotWide}
\alias{leggTilTomme_PivotWide}
\title{Utvid og organiser
Funksjon som utvider tabellen og legger til alle manglende kombinasjoner av
variablene "maned" og "legend". Deretter brukes tidyr::pivot_wider
for å få en rad per måned og en kolonne per nivå av faktoren "legend"}
\usage{
leggTilTomme_PivotWide(data, remove_NAcols = TRUE)
}
\arguments{
\item{data}{datasett med variablene maaned,legend og value}

\item{remove_NAcols}{Boolsk variabel. Remove_NA =TRUE dersom nivået "NA" skal
fjernes. FALSE, dersom vi ønsker beholde nivået "NA" og ha en kolonne for
dette nivået}
}
\value{
data.frame med en rad per nivå av maaned og en kolonne per nivå av
"legend"
}
\description{
Utvid og organiser
Funksjon som utvider tabellen og legger til alle manglende kombinasjoner av
variablene "maned" og "legend". Deretter brukes tidyr::pivot_wider
for å få en rad per måned og en kolonne per nivå av faktoren "legend"
}
\examples{
x <- data.frame(maaned = rep(c("jan", "feb", "mars", "april"), 2),
  legend = factor(x = c(rep("en",4) ,rep("to",4)),
  levels = c("en", "to", "tre")),
  value = paste0("1/", 1:8))
leggTilTomme_PivotWide(x, remove_NA = FALSE)
}
