% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner.R
\name{make_sorters}
\alias{make_sorters}
\title{Make sorters for pivot table}
\usage{
make_sorters(df)
}
\arguments{
\item{df}{ablanor table}
}
\value{
a text string containing a function that defines the sorters for
each factor-variable. This should be given in input in the getPivotTable from shiny
}
\description{
For all factor-variables, get the order of levels and use this as input
in  rpivottable.
}
\examples{
df <- data.frame(kjonn = factor(x = c("M", "K", "M", "M"),
                                levels = c("M", "K"),
                                ordered = TRUE))
s <- ablanor::make_sorters(df)
s
\dontrun{
rpivotTable::rpivotTable(df, sorters = make_sorters(df, sorters =))
}
}
