% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Ablanor Shiny Application}
\usage{
run_app(browser = FALSE, logAsJson = FALSE)
}
\value{
An object representing the Ablanor app
}
\description{
Run the Ablanor Shiny Application
}
