% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner.R
\name{string_pad}
\alias{string_pad}
\title{Pad strings with empty spaces to obtain same column width in tables}
\usage{
string_pad(string_vector)
}
\arguments{
\item{string_vector}{Vector of strings to be padded, using length of the
longest string as length}
}
\value{
Character string, all strings are of same length
}
\description{
Pad strings with empty spaces to obtain same column width in tables
}
\examples{
string_pad(string_vector = c("Ja", "Nei", "Kanskje"))
}
