% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_alder}
\alias{utlede_alder}
\title{Add variable age (alder)}
\usage{
utlede_alder(df)
}
\arguments{
\item{df}{data.frame with AblaNor data. Must contain variables
\code{birth_date} and \code{dato_pros}}
}
\value{
Returns \code{df} with one new variable: alder (= age in years)
}
\description{
Age in years at time of procedure. Calculated from variables
\code{birth_date} and \code{dato_pros}.
}
\examples{
x <- data.frame(
   birth_date = as.Date(rep("1951-12-30",3), format = "\%Y-\%m-\%d"),
   dato_pros = as.Date(c("2021-05-29", "2021-12-29", "2021-12-30"),
   format = "\%Y-\%m-\%d"))
ablanor::utlede_alder(x)
}
