% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_alder_65_75}
\alias{utlede_alder_65_75}
\title{Add variable alder_65_75}
\usage{
utlede_alder_65_75(df)
}
\arguments{
\item{df}{ablanor data.frame, must contain variable \code{alder}. See also
\link[ablanor]{utlede_alder}.}
}
\value{
Returns \code{df} with one new variable: alder_65_75
#' @details
\itemize{
\item{"<65"}
\item{"[65-75>"}
\item{">=75"}
\item{NA}
}
}
\description{
Create variable \code{alder_65_75} from \code{alder}.
}
\examples{
x <- data.frame(alder = c(NA, 60:79))
ablanor::utlede_alder_65_75(x)
}
