% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_alder_75}
\alias{utlede_alder_75}
\title{Add binary variable alder_75}
\usage{
utlede_alder_75(df)
}
\arguments{
\item{df}{ablanor data.frame, must contain variable \code{alder}. See also
\link[ablanor]{utlede_alder}.}
}
\value{
Returns \code{df} with one new variable: alder_75
}
\description{
Create binary variable \code{alder_75} from \code{alder}.
}
\examples{
x <- data.frame(alder = 73:76)
ablanor::utlede_alder_75(x)
}
