% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_bmi}
\alias{utlede_bmi}
\title{Add variable bmi_manual
Add manually calculatet value of Body Mass Index
(BMI = weight in kilograms divided by the square of height in meters).}
\usage{
utlede_bmi(df)
}
\arguments{
\item{df}{data.frame with ablanor-data. Must contain variables
\code{hoyde} and \code{vekt}.}
}
\value{
Returns \code{df} with 1 new column: \code{bmi_manual}
}
\description{
Add variable bmi_manual
Add manually calculatet value of Body Mass Index
(BMI = weight in kilograms divided by the square of height in meters).
}
\examples{
df <- data.frame(hoyde = c(150, 160, 170), vekt = c(95, 85, 65))
}
