% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_bmi_klasse}
\alias{utlede_bmi_klasse}
\title{Add varibale bmi_klasse}
\usage{
utlede_bmi_klasse(df)
}
\arguments{
\item{df}{data.frame with abalnor-data. Must contain variable
\code{bmi}.}
}
\value{
Returns \code{df} with one new column: \code{bmi_klasse}.
}
\description{
Add categories of Body Mass Index (BMI = weight in kilograms divided by
the square of height in meters).
}
\examples{
df <- data.frame(bmi = c(15, 15.2, 19, 25, 26.7, 32.1, 41.0, NA))
}
