% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_kvalitetsindikatorer.R
\name{utlede_kvalitetsindikatorer}
\alias{utlede_kvalitetsindikatorer}
\alias{indik_tamponade}
\alias{indik_avbrudd}
\title{Kvalitetsindikatorer for AblaNor}
\usage{
indik_tamponade(df)

indik_avbrudd(df)
}
\arguments{
\item{df}{data.frame med ablanor-data. Må inneholde ulike variabler for de
ulike funksjonene.  F.eks. \code{forlopstype}, \code{abla_strat_av_his} og
\code{komp_tamp} for indikatoren "Komplikasjon tamponade for AFLI uten AV knuter" .}
}
\description{
Oppdaterte versjoner av kvalitetsindikatorer. I bruk fra og med våren 2022.
}
\details{

}
\examples{
 df <- data.frame(forlopstype = c(2, 3, 4, NA, 1, 1, 1, 1),
                  abla_strat_av_his = c(NA, 1, 0, 0, 1, 0, 0, 0),
                  komp_tamp = c(rep(0, 6), 1, 1))
ablanor::indik_tamponade(df = df)
}
