% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_tidsvariabler}
\alias{utlede_tidsvariabler}
\title{Add tidsvariabler}
\usage{
utlede_tidsvariabler(df)
}
\arguments{
\item{df}{data.frame with ablanor-data. Must contain \code{dato_pros}.}
}
\value{
Returns \code{df} with 3 new columns
}
\description{
Add variables \code{aar}, \code{maaned_nr} (numerical month) and
\code{maaned} (year + month) in data set based in \code{dato_pros}.
}
\examples{
 df <- data.frame(
dato_pros = as.Date(c("2021-10-05", "1998-01-16", "2020-07-07"),
                    format = "\%Y-\%m-\%d"))
ablanor::utlede_tidsvariabler(df)
}
