% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{getRepoContributors}
\alias{getUserPubkey}
\alias{selectListPubkey}
\alias{dumpCommand}
\alias{dumpFile}
\alias{getGithub}
\alias{exportDb}
\title{Tools for exporting data}
\usage{
getGithub(what, value, .token = NULL)

selectListPubkey(pubkey)

exportDb(registryName, compress = FALSE, session)
}
\arguments{
\item{what}{Character string defining the api endpoint. Currently
one of \code{c("contributors", "keys")}.}

\item{value}{Character string specifying what to collect from the given
endpoint. For "contributors" this should be the name of the repository and
for "keys" this should be a github user name.}

\item{.token}{Character string providing av valid token that will be used if
the api call requires authentication.}

\item{pubkey}{Character vector with public keys}

\item{compress}{Logical if export data is to be compressed (using gzip).
FALSE by default.}

\item{session}{Shiny session object}

\item{repo}{Character string providing github repository name}

\item{user}{Character string with github username}
}
\value{
Some data structure with info obtained from the github api
}
\description{
Tools for exporting data
}
