% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{getRepoContributors}
\alias{getUserPubkey}
\alias{selectListPubkey}
\alias{dumpCommand}
\alias{dumpFile}
\title{Tools for exporting data}
\usage{
getRepoContributors(repo)

getUserPubkey(user)

selectListPubkey(pubkey)

dumpCommand(registryName)

dumpFile(registryName)
}
\arguments{
\item{repo}{Character string providing github repository name}

\item{user}{Character string with github username}

\item{pubkey}{Character vector with public keys}
}
\value{
Some data structure with info obtained from the github api
}
\description{
Tools for exporting data
}
