% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getDataAblanor}
\alias{getDataAblanor}
\alias{getBasereg}
\alias{getPros}
\alias{getMce}
\alias{getRand12}
\alias{getFollowupBasis}
\alias{getFollowupOneYr}
\alias{getFollowupFiveYr}
\alias{getGkv}
\alias{getProms}
\alias{getHendelse}
\alias{getPatientlist}
\alias{getFriendlycentre}
\alias{getMcepatientdata}
\alias{getBaseregPros}
\alias{getLatestEntry}
\alias{getNameReshId}
\alias{getHospitalName}
\alias{getOperatorliste}
\title{Ablanor SQL}
\usage{
getBasereg(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getPros(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getMce(singleRow, reshId = NULL, userRole, fromDate = NULL, toDate = NULL, ...)

getRand12(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupBasis(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupOneYr(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupFiveYr(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getGkv(singleRow, reshId = NULL, userRole, fromDate = NULL, toDate = NULL, ...)

getProms(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getHendelse(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFriendlycentre(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getMcepatientdata(singleRow, reshId = NULL, userRole, fromDate, toDate, ...)

getPatientlist(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregPros(
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getLatestEntry()

getNameReshId(asNamedList = FALSE, shortNames = FALSE, newNames = FALSE)

getHospitalName(reshId, shortName = FALSE, newNames = FALSE)

getOperatorliste(singleRow, reshId = NULL, ...)
}
\arguments{
\item{singleRow}{bools. TRUE only one row, for variable names. FALSE all
rows.}

\item{reshId}{Integer organization id. From login settings.}

\item{userRole}{String dummy/placeholder role. "LC" has access only
to local data (defined by reshId), "SC" has access to national data.}

\item{fromDate}{NULL default is 01-01-1900. If datadump or pivot table,
start date of calendar is used.}

\item{toDate}{NULL default is newest registration in Abalnor. If datadump or
pivot table, end date of calendar is used.}

\item{...}{Optional arguments to be passed to the function.}

\item{asNamedList}{Logical whether to return a list of named values or not.
Default is FALSE in which case a data frame containing name and id is
returned.}

\item{newNames}{boolean. TRUE uses "sykehusnavn" as defined in
'legg_til_sykehusnavn()'. Default value is FALSE, uses "sykehusnavn" from
table Friendlycentre.}

\item{shortName}{boolean. Default value FALSE and "friendlyname" is returned.
If TRUE shortname is returned.}
}
\value{
Data frame or (when multiple data sets are returned) a list of data
frames containing registry data. In case of \code{getNameReshId()} data may
also be returned as a named list of values (see Details).
}
\description{
Get tables from database usig SQL. Notice, only tables when a
procedure-date exists!
}
\details{
\code{getNameReshId()} returns a mapping of organization name and id in the
form of columns named \emph{name} and \emph{id}. Optionally this function
can also return a list of named values (ids), \emph{e.g.} for use in shiny
selection lists.
}
