% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKodebokData.R
\name{getKodebokData}
\alias{getKodebokData}
\title{Behandling av kodeboken (klokeboken) før bruk}
\usage{
getKodebokData()
}
\value{
data.frame med kodeboken til AblanNor. Nøkkelkolonner:
\code{fysisk_feltnavn} (variabel-navn)
\code{type} (variabeltype), \code{listeverdier} (numeriske verdier for
Listevariabler) og \code{listetekst} (tekst-verdi for Listevariabler).
Hver variabel har en rad i \code{kodeboken}. For variabler av \code{type}
"Listevariabel"  inneholder kodeboken en rad per listeverdi.
}
\description{
Les inn kodeboken
}
