% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKodebokData.R
\name{getKodebokData}
\alias{getKodebokData}
\title{Behandling av kodeboken (klokeboken) før bruk}
\usage{
getKodebokData(registryName, singleRow = FALSE, tekstVars = FALSE, ...)
}
\arguments{
\item{registryName}{Dersom verdien "test_ablanor_lokalt" leser vi inn
lokal RData-fil. Ellers er det SQL spørring}

\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{tekstVars}{legge til tekstvariabler hentet fra kodebok for kategoriske
variabler eller fra tabeller med tekst/labels for kategoriske variabler}

\item{...}{Optional arguments to be passed to the function}
}
\value{
data.frame med kolonnene \code{fysisk_feltnavn} (variabel-navn)
\code{type} (variabeltype), \code{listeverdier} (numeriske verdier for
Listevariabler) og \code{listetekst} (tekst-verdi for Listevariabler).
Hver variabel har en rad i \code{kodeboken}. For variabler av \code{type}
"Listevariabel"  inneholder kodeboken en rad per listeverdi.
}
\description{
Les inn kodeboken og velge ut enkelte variabler
}
