% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPivotDataSet.R
\name{getPivotDataSet}
\alias{getPivotDataSet}
\title{Get selected data set for Pivot Table  (Utforsker)}
\usage{
getPivotDataSet(
  setId = "",
  registryName,
  session,
  singleRow = FALSE,
  reshId = NULL,
  allData = FALSE
)
}
\arguments{
\item{setId}{String providing name of data set to be returned. One of the
values "pros_patient" containing procedure and patient data or "rand12"
containing data from rand12 questionnaire.}

\item{registryName}{String providing registry name.}

\item{session}{List shiny session object.}

\item{singleRow}{Logical defining if only one row is to be returned.}

\item{reshId}{Integer organization id}

\item{allData}{Logical if global data set is to be returned. When FALSE
(default) data will be filtered by \code{reshId}}
}
\value{
data frame
}
\description{
Some tables are returned as they are, with only raw data. Other tables are
treated (importation of variables from other tables, generation of new
variables) before the data frame is returned.
}
\examples{
\dontrun{d_ablanor <- getPivotDataSet(setId = "pros_patient")
}
}
