% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPivotDataSet.R
\name{getPivotDataSet}
\alias{getPivotDataSet}
\title{Get selected data set for Pivot Table  (Utforsker)}
\usage{
getPivotDataSet(
  setId = "",
  session,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL
)
}
\arguments{
\item{setId}{String providing name of data set to be returned. One of the
values "pros_patient" containing procedure and patient data or "rand12"
containing data from rand12 questionnaire.}

\item{session}{List shiny session object.}

\item{singleRow}{Logical defining if only one row is to be returned.}

\item{reshId}{Integer organization id}

\item{userRole}{String dummy/placeholder role. "LC" has access only
to local data (defined by reshId), "SC" has access to national data.}
}
\value{
data frame
}
\description{
Some tables are returned as they are, with only raw data. Other tables are
treated (importation of variables from other tables, generation of new
variables) before the data frame is returned.
}
\examples{
\dontrun{d_ablanor <- getPivotDataSet(setId = "pros_patient")
}
}
