% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrepData.R
\name{getPrepDataAblanor}
\alias{getPrepDataAblanor}
\alias{getProsData}
\alias{getBaseregData}
\alias{getRand12Data}
\alias{getFollowupData}
\title{Data managment on tables}
\usage{
getProsData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getRand12Data(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupData(registryName, singleRow = FALSE, reshId = NULL, userRole, ...)
}
\arguments{
\item{registryName}{"ablanor"}

\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{reshId}{Integer organization id. From login settings.}

\item{userRole}{String dummy/placeholder role. "LC" has access only
to local data (defined by reshId), "SC" has access to national data.}

\item{fromDate}{NULL default is newest registration in Abalnor.
Can be set to end date if calendar is chosen (downoad/pivot table) or any
other end-date (report).}
}
\value{
data.frame med rad per forløp og kolonner for variabler
}
\description{
Load data and apply data-management operations. Tables can be used in
\emph{pivot-tables} or monthly reports. Notice, only tables when a
procedure-date exists!
}
