% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrepData.R
\name{getPrepDataAblanor}
\alias{getPrepDataAblanor}
\alias{getBaseregData}
\alias{getProsData}
\alias{getMceData}
\alias{getRand12Data}
\alias{getFollowupData}
\alias{getGkvData}
\alias{getBaseregProsData}
\title{Data managment on tables}
\usage{
getBaseregData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getProsData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getMceData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getRand12Data(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getGkvData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupData(registryName, singleRow = FALSE, reshId = NULL, userRole, ...)
}
\arguments{
\item{registryName}{"ablanor"}

\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{reshId}{Integer organization id. From login settings.}

\item{userRole}{String dummy/placeholder role. "LC" has access only
to local data (defined by reshId), "SC" has access to national data.}

\item{fromDate}{NULL default is 01-01-1900. If datadump or pivot table,
start date of calendar is used.}

\item{toDate}{NULL default is newest registration in Abalnor. If datadump or
pivot table, end date of calendar is used.}
}
\value{
data.frame med rad per forløp og kolonner for variabler
}
\description{
Load data and apply data-management operations. Tables can be used in
\emph{pivot-tables} or monthly reports. Notice, only tables when a
procedure-date exists!
}
