% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProsPatientData.R
\name{getProsPatientData}
\alias{getProsPatientData}
\title{Hent merget datasett
Hent datasett fra basis-skjema, prosedyreskjema, mce, og pasientskjema. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til utledete
variabler og nye variabler.}
\usage{
getProsPatientData(registryName, singleRow = FALSE, tekstVars = FALSE)
}
\arguments{
\item{registryName}{Dersom verdien "test_ablanor_lokalt" leser vi inn
lokal RData-fil. Ellers er det SQL spørring}

\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{tekstVars}{legge til tekstvariabler hentet fra kodebok for kategoriske
variabler eller fra tabeller med tekst/labels for kategoriske variabler}
}
\value{
data.frame med rad per forløp og kolonner for variabler
}
\description{
Hent merget datasett
Hent datasett fra basis-skjema, prosedyreskjema, mce, og pasientskjema. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til utledete
variabler og nye variabler.
}
