% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProsPatientData.R
\name{getProsPatientData}
\alias{getProsPatientData}
\title{Hent merget datasett
Hent datasett fra basis-skjema, prosedyreskjema, mce, og pasientskjema. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til utledete
variabler og nye variabler.}
\usage{
getProsPatientData(
  registryName,
  singleRow = FALSE,
  reshId = NULL,
  userRole,
  fromDate,
  toDate,
  ...
)
}
\arguments{
\item{registryName}{Dersom verdien "test_ablanor_lokalt" leser vi inn
lokal RData-fil. Ellers er det SQL spørring}

\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{reshId}{Integer organization id}

\item{userRole}{String dummy/placeholder role. "LC" has access only
to local data (defined by reshId), "SC" has access to national data.}

\item{fromDate}{Character string of format YYYY-MM-DD with start date. Value
NULL if no filter on date.}

\item{toDate}{Character string of format YYYY-MM-DD with end date. Value
NULL if no filter on date.}

\item{...}{Optional arguments to be passed to the function}
}
\value{
data.frame med rad per forløp og kolonner for variabler
}
\description{
Hent merget datasett
Hent datasett fra basis-skjema, prosedyreskjema, mce, og pasientskjema. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til utledete
variabler og nye variabler.
}
