% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRand12Data.R
\name{getRand12Data}
\alias{getRand12Data}
\title{Hent merget datasett
Hent datasett fra prosedyreskjema,og rand12. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til evt utledete
variabler og nye variabler.}
\usage{
getRand12Data(registryName, singleRow = FALSE, ...)
}
\arguments{
\item{registryName}{Dersom verdien "test_ablanor_lokalt" leser vi inn
lokal RData-fil. Ellers er det SQL spørring}

\item{singleRow}{bools. TRUE bare metadata, FALSE hele datasettet}

\item{...}{Optional arguments to be passed to the function}
}
\value{
data.frame med rad per forløp og kolonner for variabler
}
\description{
Hent merget datasett
Hent datasett fra prosedyreskjema,og rand12. Merge
sammen, filtrere på forløp der prosedyre er utført. Legge til evt utledete
variabler og nye variabler.
}
