% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryDataAblanor}
\alias{queryDataAblanor}
\alias{queryProm0}
\alias{queryRand12_0}
\title{Ablanor SQL}
\usage{
queryProm0()

queryRand12_0()
}
\value{
Data frame or (when multiple data sets are returned) a list of data
frames containing registry data. In case of \code{getNameReshId()} data may
also be returned as a named list of values (see Details).
}
\description{
Get tables from database usig SQL. Notice, only tables when a
procedure-date exists!
}
