% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_CHA2DS2VASc}
\alias{utlede_CHA2DS2VASc}
\title{CHA2DS2VASc}
\usage{
utlede_CHA2DS2VASc(df)
}
\arguments{
\item{df}{data.frame with Ablanor-data. Must contain variables
\code{hjertesvikt}, \code{hypertoni} and \code{diabetes},
\code{tia_slag}, \code{alder} and \code{gender}.}
}
\value{
Returns \code{df} with 1 new column
}
\description{
Add variable \code{CHA2DS2VASc} score from 0 to 9.
#'
}
\details{
\itemize{
\item{hjertesvikt:}{ score +1}
\item{hypertoni:}{ score +1}
\item{karsykdom:}{ score +1}
\item{diabetes:}{ score +1}
\item{tia_slag:}{ score +2}
\item{gender(kvinne:)}{ score +1}
\item{alder(>=65; <75):}{ score +1}
\item{alder(>=75):}{ score +2}

}
}
\examples{
  df <- data.frame(
    hjertesvikt = c(NA, 2, rep(1, 8)),
    hypertoni = c(NA, 1, NA, 0, 0, 9, 9, 1, 1, 1),
    karsykdom = c(0, 1, 1, 1, 1, 0, 0, NA, NA, 9),
    diabetes = c(NA, 9, 1, 9, 1, 1, NA, 1, 0, 9),
    tia_slag = c(NA, 1, 1, 9, 1, 1, 0, 1, 0, 0),
    alder = c(18, 25, 64, 65, 70, 74, 75, 76, 80, 105),
    gender = c(rep(1, 4), NA, NA, rep(2, 4)))

 ablanor::utlede_CHA2DS2VASc(df)
}
