% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_kvalitetsindikatorer.R
\name{utlede_kvalitetsindikatorer}
\alias{utlede_kvalitetsindikatorer}
\alias{indik_tamponade}
\alias{indik_avbrudd}
\title{Kvalitetsindikatorer for AblaNor}
\usage{
indik_tamponade(df)

indik_avbrudd(df)
}
\arguments{
\item{df}{data.frame med ablanor-data. Må inneholde ulike variabler for de
ulike funksjonene.  F.eks. \code{forlopstype}, \code{abla_strat_av_his} og
\code{komp_tamp} for indikatoren "Komplikasjon tamponade for AFLI uten AV knuter" .}
}
\description{
Oppdaterte versjoner av kvalitetsindikatorer. I bruk fra og med våren 2022.
}
\details{
For hver av kvalitetsindikatorene, legge til en variabel for datagrunnlag
(suffix 'data') og en for indikatoren ja/nei.

\strong{Tamponade i forbindelse med prosedyren}
\code{indik_tamponade()}
\itemize{
\item nevneren \code{indik_tamponade_data} (datagrunnlag) har verdien \emph{ja} dersom
forløpstype er AFLI (\code{forlopstype} = 1)
uten AV-knuter (\code{abla_strat_av_his} = 0).
\item telleren \code{indik_tamponade} har verdien \emph{ja} dersom
\code{indik_tamp_data} = \emph{ja} og \code{komp_tamp} = 1,
verdien \emph{nei} dersom \code{indik_tamp_data} = \emph{ja} og \code{komp_tamp} = 0,
verdien \emph{manglende} dersom \code{indik_tamp_data} = \emph{ja} og
\code{komp_tamp} er manglende,
og verdien \emph{NA} dersom forløpet ikke er i datagrunnlaget (\code{indik_tamp_data} = \emph{nei}).
}

\strong{Prosedyreavbrudd på grunn av tekniske problemer eller komplikasjoner}
\code{indik_avbrudd()}
\itemize{
\item nevneren \code{indik_avbrudd_data} (datagrunnlag) har
verdien \emph{ja} dersom forløpstype er AFLI (\code{forlopstype} = 1)
uten AV-knuter (\code{abla_strat_av_his} = 0) og dersom pasienten er abladert
(\code{abla_strat_ingen} er ikke manglende).
\item telleren \code{indik_avbrudd} har verdien \emph{ja} dersom
\code{indik_avbrudd_data} = \emph{ja} og
\code{abla_strat_ingen_arsak} = 4(tekniske problemer) eller 5(Komplikasjon),
verdien \emph{nei} dersom \code{indik_avbrudd_data} = \emph{ja} og
ingen avbrudd eller avbrudd av andre årsaker.
Verdien \emph{NA} dersom forløpet ikke er i datagrunnlaget.
}
}
\examples{
 df <- data.frame(forlopstype = c(2, 3, 4, NA, 1, 1, 1, 1),
                  abla_strat_av_his = c(NA, 1, 0, 0, 1, 0, 0, 0),
                  komp_tamp = c(rep(0, 6), 1, 1))
ablanor::indik_tamponade(df = df)
}
